Simulates CPASS data for a single cycle.

This function simulates DRSP data for a single cycle

Usage

simulate_cycle(cycle_nb, baseline, pmdd_intensity, items, missingness_rate)

Arguments

cycle_nb

(positive integer) The cycle number.

baseline

(positive value in 1-6) Baseline level for severity of problems for all items.

pmdd_intensity

change from baseline during the pre-menstrual phase.

items

(positive integers in 1-24) items for which there is a variation in the pre-menstrual phase.

missingness_rate

(positive value in 0-0.9) The fraction of missing days.

Value

a data.frame providing the DRSP scores for each cycle day (within the pre- and post- menstrual phase) and item.

See also