Supported data format for the as_cpass_data function.

This function prints (or return) a list of the format currently supported by the cpass package.

supported_data_format(as_list = FALSE)

Arguments

as_list

(optional, default = FALSE) logical specifying if the supported format should be printed to the console or returned as a list object.

Value

a list if as_list = TRUE.

Examples

supported_data_format()
#> 
#> 1) CPASS FORMAT
#>      columns:  cycle, day 
#>      description:
#>      *  cycle is the cycle number 
#>      *  day is counted forward (1,2,3,...) for the post-menses phase;
#>              backward (-1,-2,...) for the pre-menses one. 
#> 
#> 
#> 2) FIRST DAY
#>      columns:  date, is_cycle_start 
#>      description:
#>      *  `date` is an actual date
#>           (use 'as.Date' to convert character strings to dates) 
#>      *  `is_cycle_start` is a logical which is TRUE
#>           if a menstrual cycle starts on the specified date. 
#> 
#> 
#> 3) BLEEDING
#>      columns:  date, bleeding 
#>      description:
#>      *  `date` is an actual date
#>           (use 'as.Date' to convert character strings to dates) 
#>      *  `bleeding` is a character describing vaginal bleeding.
#>           It can take the following values:
#>           'none','spotting','light', 'medium', 'heavy'.
#>           Missing values (NAs) are accepted. 
#>